Consolidate agentcomms.go into machine_mgr.go
